refactor: DIViewModelFactory, ShoppingApplication 패키지 변경
