+++ app/src/main/java/woowacourse/shopping/common/DIViewModelFactory.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.common
 
 import android.app.Application
 import androidx.lifecycle.ViewModel
