+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -7,11 +7,10 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.R
+import woowacourse.shopping.ShoppingApplication
+import woowacourse.shopping.common.DIViewModelFactory
 import woowacourse.shopping.databinding.ActivityCartBinding
-import woowacourse.shopping.ui.DIViewModelFactory
-import woowacourse.shopping.ui.ShoppingApplication
 
 class CartActivity : AppCompatActivity() {
     private val binding by lazy { ActivityCartBinding.inflate(layoutInflater) }
